pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastembed>=0.7.4",
    "ipykernel>=7.1.0",
    "llama-index>=0.14.10",
    "llama-index-core>=0.14.10",
    "llama-index-embeddings-fastembed>=0.5.0",
    "llama-index-llms-openai>=0.6.12",
    "llama-index-vector-stores-qdrant>=0.9.0",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "streamlit>=1.52.2",
]
